move type writer header to utility folder
