+++ swiftwinrt/code_writers/class_writers.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "types.h"
-#include "../type_writers.h"
+#include "utility/type_writers.h"
 namespace swiftwinrt
 {
     void write_class(writer& w, class_type const& type);
