+++ swiftwinrt/code_writers/common_writers.h
@@ -1,8 +1,8 @@
 #pragma once
 #include "utility/metadata_helpers.h"
 #include "utility/swift_codegen_utils.h"
-#include "../type_writers.h"
-#include "type_writers.h"
+#include "code_writers/type_writers.h"
+#include "utility/type_writers.h"
 
 namespace swiftwinrt
 {
