+++ swiftwinrt/code_writers/delegate_writers.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "types.h"
-#include "../type_writers.h"
+#include "utility/type_writers.h"
 namespace swiftwinrt
 {
     void write_delegate(writer& w, delegate_type const& type);
