+++ swiftwinrt/code_writers/enum_writers.cpp
@@ -1,6 +1,6 @@
 #include "utility/metadata_cache.h"
 #include "common_writers.h"
-#include "type_writers.h"
+#include "utility/type_writers.h"
 #include "can_write.h"
 #include "enum_writers.h"
 #include "utility/metadata_helpers.h"
