+++ swiftwinrt/code_writers/enum_writers.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "types.h"
-#include "../type_writers.h"
+#include "utility/type_writers.h"
 namespace swiftwinrt
 {
     void write_enum_def(writer& w, enum_type const& type);
