+++ swiftwinrt/code_writers/type_writers.cpp
@@ -1,8 +1,8 @@
-#include "../pch.h"
-#include "../types.h"
-#include "../type_writers.h"
+#include "pch.h"
+#include "types.h"
+#include "code_writers/type_writers.h"
 
-#include "type_writers.h"
+#include "utility/type_writers.h"
 #include "utility/separator.h"
 #include "utility/swift_codegen_utils.h"
