+++ swiftwinrt/types/typedef_base.cpp
@@ -4,7 +4,7 @@
 
 #include "utility/metadata_helpers.h"
 #include "utility/type_helpers.h"
-#include "type_writers.h"
+#include "utility/type_writers.h"
 #include "utility/versioning.h"
 
 using namespace std::literals;
