+++ swiftwinrt/utility/separator.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "type_writers.h"
+#include "utility/type_writers.h"
 
 namespace swiftwinrt {
     struct separator
