+++ swiftwinrt/utility/type_writers.h
